+++ src/mocks/handlers/milestone.ts
@@ -2,42 +2,8 @@
 import { rest } from 'msw';
 import { ContentTypes, MilestoneTypes } from '@/api/issue/types';
 import { issueTable } from '@/mocks/tables/issue';
-import { MilestoneListTypes } from '@/api/milestone';
+import { MILESTONE_TABLE } from '@/mocks/tables/milestone';
 import { ERROR_CODE } from '@/api/constants';
-
-export const MILESTONE_TABLE: MilestoneListTypes = {
-  openedMilestones: [
-    {
-      id: 0,
-      title: '마일스톤 1',
-      description: null,
-      dueDate: null,
-      closed: false,
-      openIssueCount: 3,
-      closedIssueCount: 7,
-    },
-    {
-      id: 2,
-      title: '마일스톤 3',
-      description: '열린 마일스톤에 대한 설명',
-      dueDate: '2022-08-28',
-      closed: false,
-      openIssueCount: 5,
-      closedIssueCount: 5,
-    },
-  ],
-  closedMilestones: [
-    {
-      id: 1,
-      title: '마일스톤 2',
-      description: '닫힌 마일스톤에 대한 설명',
-      dueDate: null,
-      closed: true,
-      openIssueCount: 16,
-      closedIssueCount: 13,
-    },
-  ],
-};
 
 export const findMilestoneHelper = (id: number) =>
   Object.values(MILESTONE_TABLE)
